jitx-py/sharp-sc62015-interposer/pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0,<2.0"]
build-backend = "hatchling.build"

[project]
name = "sharp-sc62015-interposer-py"
version = "0.1.0"
description = "Python JITX port scaffold for the retrobus sharp-sc62015-interposer board"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "jitx>=4.0.0,<5",
  "jitxlib-parts>=1.0.0,<2",
  "jitxlib-standard>=4.0.0,<5",
  "jitxexamples-components>=1.0.0,<2",
  "retrobus-jitx-parts",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.optional-dependencies]
dev = [
  "ruff",
  "ty",
]

[tool.uv.sources]
retrobus-jitx-parts = { path = "../retrobus-jitx-parts", editable = true }

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["E501"]
